+++ app/templates/components/modal-dialog-tethered-structure.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-modal-dialog/templates/components/modal-dialog-tethered-structure';