+++ app/templates/components/modal-dialog.js
@@ -1,2 +1,0 @@
-import template from 'ember-modal-dialog/templates/components/modal-dialog';
-export default template;